+++ CHANGELOG.md
@@ -14,3 +14,6 @@ Should match the commit history, or at least give you an idea of what changed ha
 0. Fixed a bug in `hc_array_resize_generic()` that had USE AFTER FREE. Still need to rewrite this function... Eventually.
 0. Build script now has `--test` command that runs a specific example with all precautions.
 0. Added `hc_` prefix to where it was missing.
+0. Removed old `./examples`.
+0. Added missing `--nerd` flag response for `./build.c`.
+0. Updated `array.c` to have use of `hc_array_unordered_remove()`.